Refactor http3 part 1.
