+++ neqo-http3/src/lib.rs
@@ -7,10 +7,13 @@
 #![deny(warnings)]
 
 mod client_events;
-pub mod connection;
+mod connection;
+pub mod connection_client;
+pub mod connection_server;
 mod control_stream_local;
 mod control_stream_remote;
 pub mod hframe;
+mod server_events;
 mod stream_type_reader;
 mod transaction_client;
 pub mod transaction_server;
@@ -22,7 +25,9 @@ pub use neqo_transport::Output;
 use self::hframe::HFrameType;
 
 pub use client_events::Http3ClientEvent;
-pub use connection::{Http3Connection, Http3State};
+pub use connection::Http3State;
+pub use connection_client::Http3Client;
+pub use connection_server::Http3Server;
 pub use neqo_qpack::Header;
 pub use transaction_server::TransactionServer;
