+++ test-fixture/Cargo.toml
@@ -7,5 +7,6 @@ edition = "2018"
 [dependencies]
 neqo-common = { path = "../neqo-common" }
 neqo-crypto = { path = "../neqo-crypto" }
+neqo-http3 = { path = "../neqo-http3" }
 neqo-transport = { path = "../neqo-transport" }
 log = "0.4.0"
